extract field helpers to fields.ts
