+++ src/services/tagging/idTaggingScheme.ts
@@ -1,8 +1,8 @@
-import { Feature } from '../types';
-import { getFieldTranslation, getPresetTranslation } from './translations';
-import { getPresetForFeature } from './presets';
-import { Preset } from './types/Presets';
-import { fields, presets } from './data';
+import { Feature } from "../types";
+import { getFieldTranslation, getPresetTranslation } from "./translations";
+import { getPresetForFeature } from "./presets";
+import { fields } from "./data";
+import { getAllFieldKeys, getValueForField } from "./fields";
 
 // TODO move to shared place
 const featuredKeys = [
@@ -15,57 +15,6 @@ const featuredKeys = [
   'opening_hours',
   'description',
 ];
-
-// links like {shop}, are recursively resolved to their fields
-const getResolvedFields = (fieldKeys: string[]): string[] =>
-  fieldKeys.flatMap((key) => {
-    if (key.match(/^{.*}$/)) {
-      const presetKey = key.substr(1, key.length - 2);
-      return getResolvedFields(presets[presetKey].fields); // TODO does "{shop}" links to preset's fields or moreFields?
-    }
-    return key;
-  });
-
-const getResolvedFieldsWithParents = (
-  preset: Preset,
-  fieldType: 'fields' | 'moreFields',
-): string[] => {
-  const parts = preset.presetKey.split('/');
-
-  if (parts.length > 1) {
-    const parentKey = parts.slice(0, parts.length - 1).join('/');
-    const parentPreset = presets[parentKey];
-    if (parentPreset) {
-      return [
-        ...getResolvedFieldsWithParents(parentPreset, fieldType),
-        ...(preset[fieldType] ?? []),
-      ];
-    }
-  }
-
-  return preset[fieldType] ?? [];
-};
-
-const getAllFieldKeys = (preset: Preset) => {
-  const allFieldKeys = [
-    ...getResolvedFields(getResolvedFieldsWithParents(preset, 'fields')),
-    ...getResolvedFields(getResolvedFieldsWithParents(preset, 'moreFields')),
-  ];
-
-  // @ts-ignore
-  return [...new Set(allFieldKeys)];
-};
-
-const getValueForField = (field, fieldTranslation, value: string) => {
-  if (field.type === 'semiCombo') {
-    return value
-      .split(';')
-      .map((v) => fieldTranslation?.options?.[v] ?? v)
-      .join(',\n');
-  }
-
-  return fieldTranslation?.options?.[value] ?? value;
-};
 
 export const getSchemaForFeature = (feature: Feature) => {
   const preset = getPresetForFeature(feature);
